1.3.1: HoundSploit now can automatically update the database

+++ setup.py
@@ -1,7 +1,13 @@
 import os
+from searcher.engine.updates import install_exploitdb_update
 
 
 def main():
+    if not os.path.isfile('houndsploit.py'):
+        print("Change the current working directory to the directory of \'houndsploit.py\'"
+              " for executing \'setup.py\'.")
+        exit(0)
+
     os.system('pip install --upgrade pip')
 
     try:
@@ -33,6 +39,8 @@ def main():
         import requests
     except ImportError:
         os.system('pip install requests')
+
+    install_exploitdb_update()
 
 
 if __name__ == "__main__":
